Refactor user service and router to improve functionality and organization
- Updated users.service.ts to streamline user data retrieval and processing functions.
- Enhanced searchOrSummary, getUser, getProgression, getPersonalBests, getRank, and compare functions for better performance and clarity.
- Integrated users.service into users.ts for cleaner API route handling.
- Improved response structure and error handling in user-related endpoints.
- Added pagination support in search results and refined data formatting for competition results.
